Update gateway-config.md to clarify value format
The description currently attached to TYK_GW_SERCURITY_CERTIFICATES_UPSTREAM makes it appear that a JSON string should be set as the environmental variable to override the `security.certificatges.upstream` map in `tyk.conf`. However, using this format results in Tyk totally ignoring the specified upstream certificates. I updated the description to make it clear that the JSON map is used in `tyk.conf` while the environmental variable is set via a comma separated list of `host:cert` statements.